backend/requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# HRPulse Backend — Python Dependencies
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

--extra-index-url https://download.pytorch.org/whl/cpu

# ── Web Framework ─────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator>=2.1.0

# ── Database ──────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
greenlet==3.0.3

# ── Authentication ────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Task Queue ────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4

# ── ML / DL ──────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
torch==2.3.0
transformers==4.40.2
mlflow==2.12.1
joblib==1.4.2

# ── GenAI & Agents ───────────────────────────
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langgraph==0.0.53
llama-index==0.10.43
llama-index-core==0.10.43
llama-index-readers-file==0.1.23
faiss-cpu==1.8.0
sentence-transformers==2.7.0

# ── Document Processing ──────────────────────
pypdf2==3.0.1

# ── Utilities ─────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
